Database: fixed some broken CardAbility in db/seeds.rb

+++ db/migrate/20250703173632_add_claiming_to_game_board_tiles.rb
@@ -1,9 +1,11 @@
 class AddClaimingToGameBoardTiles < ActiveRecord::Migration[7.2]
   def up
-    add_column_unless_exists :game_board_tiles, :claiming_user_id, :integer, null: true
-    add_column_unless_exists :game_board_tiles, :claimed_at, :integer, null: true
+    if table_exists?(:game_board_tiles)
+      add_column_unless_exists :game_board_tiles, :claiming_user_id, :integer, null: true
+      add_column_unless_exists :game_board_tiles, :claimed_at, :integer, null: true
 
-    add_index :game_board_tiles, :claiming_user_id unless index_exists?(:game_board_tiles, :claiming_user_id)
+      add_index :game_board_tiles, :claiming_user_id unless index_exists?(:game_board_tiles, :claiming_user_id)
+    end
 
     drop_table_if_exists :board_tiles
     create_table :board_tiles do |t|
@@ -19,10 +21,12 @@ class AddClaimingToGameBoardTiles < ActiveRecord::Migration[7.2]
   end
 
   def down
-    remove_index :game_board_tiles, :claiming_user_id if index_exists?(:game_board_tiles, :claiming_user_id)
+    if table_exists?(:game_board_tiles)
+      remove_index :game_board_tiles, :claiming_user_id if index_exists?(:game_board_tiles, :claiming_user_id)
 
-    remove_column_if_exists :game_board_tiles, :claiming_user_id
-    remove_column_if_exists :game_board_tiles, :claimed_at
+      remove_column_if_exists :game_board_tiles, :claiming_user_id
+      remove_column_if_exists :game_board_tiles, :claimed_at
+    end
 
     drop_table_if_exists :board_tiles
   end
